% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{tk_attributes}
\alias{tk_attributes}
\title{Get list of attributes for a given entity}
\usage{
tk_attributes(entity)
}
\arguments{
\item{entity}{The entity to get attributes for}
}
\value{
A character vector with the entities
}
\description{
Get list of attributes for a given entity
}
