% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata.R
\name{tk_count}
\alias{tk_count}
\title{Count the number of results for a query}
\usage{
tk_count(tk_url)
}
\arguments{
\item{tk_url}{A tk_url_builder object}
}
\value{
a number
}
\description{
Count the number of results for a query
}
