% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata.R
\name{tk_entity}
\alias{tk_entity}
\title{Select entity}
\usage{
tk_entity(entity)
}
\arguments{
\item{entity}{}
}
\value{
A tk_url object
}
\description{
See the [Information model](https://opendata.tweedekamer.nl/documentatie/informatiemodel) for valid entities.
}
