% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata.R
\name{tk_filter}
\alias{tk_filter}
\title{Add a filter}
\usage{
tk_filter(tk_url, ...)
}
\arguments{
\item{tk_url}{A tk_url_builder object (the output of tk_entity)}

\item{...}{Filters to apply. Use by_string, by_date, by_number, by_boolean, filter_and, filter_or}
}
\description{
Add a filter
}
\examples{
tk_entity('persoon') \%>\%
  tk_filter(by_string('geslacht', 'eq', 'vrouw')) \%>\%
  tk_count()

tk_entity('persoon') \%>\%
  tk_filter(
    by_string('geslacht', 'eq', 'vrouw'),
    filter_or(
      by_date('geboortedatum', 'lt', '1970-01-01T23:00:00Z'),
      by_date('geboortedatum', 'gt', '1980-01-01')
    )
  ) \%>\%
  tk_count()
}
