% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata.R
\name{tk_list}
\alias{tk_list}
\title{Get list of results for query}
\usage{
tk_list(tk_url, top = NULL, skip = NULL)
}
\arguments{
\item{tk_url}{A tk_url object}

\item{top}{Maximum number of results to return. Maximum is 250.}

\item{skip}{Number of results to skip.}
}
\value{
A tibble with the results
}
\description{
Get list of results for query
}
