% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata.R
\name{tk_string_filter}
\alias{tk_string_filter}
\title{Filter on string value}
\usage{
tk_string_filter(field, operator, value)
}
\arguments{
\item{field}{Field to filter on}

\item{operator}{Operator to use. One of 'eq', 'contains', 'startswith', 'endswith'}

\item{value}{string value}
}
\description{
Filter on string value
}
