% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata.R
\name{tk_filter}
\alias{tk_filter}
\title{Add a filter}
\usage{
tk_filter(tk_url, ...)
}
\arguments{
\item{tk_url}{A tk_url object}

\item{...}{Filters to apply. Use tk_string_filter, tk_date_filter, tk_number_filter, tk_boolean_filter, tk_and, tk_or}
}
\description{
Add a filter
}
