% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata.R
\name{tk_list}
\alias{tk_list}
\title{Get list of results for query}
\usage{
tk_list(
  tk_url,
  limit = Inf,
  skip = NULL,
  meta = "minimal",
  expand_notnull = TRUE
)
}
\arguments{
\item{tk_url}{A tk_url_builder object (the output of tk_entity)}

\item{limit}{Number of results to return.}

\item{skip}{Number of results to skip.}

\item{meta}{Either 'none', 'minimal' or 'full'}
}
\value{
A tibble with the results
}
\description{
Get list of results for query
}
